+++ src/main/java/com/github/merchantpug/apugli/entity/feature/EntityTextureOverlayFeatureRenderer.java
@@ -3,6 +3,8 @@ package com.github.merchantpug.apugli.entity.feature;
 import com.github.merchantpug.apugli.util.TextureUtil;
 import io.github.apace100.apoli.component.PowerHolderComponent;
 import com.github.merchantpug.apugli.power.EntityTextureOverlayPower;
+import io.github.apace100.apoli.power.ModelColorPower;
+import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.OverlayTexture;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexConsumerProvider;
@@ -10,7 +12,10 @@ import net.minecraft.client.render.entity.feature.FeatureRenderer;
 import net.minecraft.client.render.entity.feature.FeatureRendererContext;
 import net.minecraft.client.render.entity.model.EntityModel;
 import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
+
+import java.util.List;
 
 public class EntityTextureOverlayFeatureRenderer<T extends LivingEntity, M extends EntityModel<T>> extends FeatureRenderer<T, M> {
 
@@ -20,14 +25,42 @@ public class EntityTextureOverlayFeatureRenderer<T extends LivingEntity, M exten
 
     @Override
     public void render(MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, LivingEntity entity, float limbAngle, float limbDistance, float tickDelta, float animationProgress, float headYaw, float headPitch) {
+        if (entity.isInvisible() && !MinecraftClient.getInstance().hasOutline(entity)) return;
         matrices.push();
         EntityModel<T> entityModel = this.getContextModel();
         PowerHolderComponent.getPowers(entity, EntityTextureOverlayPower.class).forEach(power -> {
+            float red = 1.0F;
+            float green = 1.0F;
+            float blue = 1.0F;
+            float alpha = 1.0F;
+
+            if (power.renderingPowerCompatible) {
+                List<ModelColorPower> modelColorPowers = PowerHolderComponent.getPowers(entity, ModelColorPower.class);
+                if (modelColorPowers.size() > 0) {
+                    red = modelColorPowers.stream().map(ModelColorPower::getRed).reduce((a, b) -> a * b).get();
+                    green = modelColorPowers.stream().map(ModelColorPower::getGreen).reduce((a, b) -> a * b).get();
+                    blue = modelColorPowers.stream().map(ModelColorPower::getBlue).reduce((a, c) -> a * c).get();
+                    alpha = modelColorPowers.stream().map(ModelColorPower::getAlpha).min(Float::compare).get();
+                }
+            }
+
+            RenderLayer renderLayer;
             if (power.textureUrl != null) {
                 TextureUtil.registerEntityTextureOverlayTexture(power.getUrlTextureIdentifier(), power.textureUrl);
-                entityModel.render(matrices, vertexConsumers.getBuffer(RenderLayer.getEntityCutoutNoCull(power.getUrlTextureIdentifier())), light, OverlayTexture.DEFAULT_UV, 1.0F, 1.0F, 1.0F, 1.0F);
+
+                renderLayer = alpha == 1.0 ? RenderLayer.getEntityCutoutNoCull(power.getUrlTextureIdentifier()) : RenderLayer.getEntityTranslucent(power.getUrlTextureIdentifier());
+                if (MinecraftClient.getInstance().hasOutline(entity) && entity.isInvisible()) {
+                    renderLayer = RenderLayer.getOutline(power.getUrlTextureIdentifier());
+                }
+
+                entityModel.render(matrices, vertexConsumers.getBuffer(renderLayer), light, OverlayTexture.DEFAULT_UV, red, green, blue, alpha);
             } else if (power.textureLocation != null) {
-                entityModel.render(matrices, vertexConsumers.getBuffer(RenderLayer.getEntityCutoutNoCull(power.textureLocation)), light, OverlayTexture.DEFAULT_UV, 1.0F, 1.0F, 1.0F, 1.0F);
+                renderLayer = alpha == 1.0 ? RenderLayer.getEntityCutoutNoCull(power.textureLocation) : RenderLayer.getEntityTranslucent(power.textureLocation);
+                if (MinecraftClient.getInstance().hasOutline(entity) && entity.isInvisible()) {
+                    renderLayer = RenderLayer.getOutline(power.textureLocation);
+                }
+
+                entityModel.render(matrices, vertexConsumers.getBuffer(renderLayer), light, OverlayTexture.DEFAULT_UV, red, green, blue, alpha);
             }
         });
         matrices.pop();