+++ src/main/java/com/github/merchantpug/apugli/power/EntityTextureOverlayPower.java
@@ -14,6 +14,7 @@ public class EntityTextureOverlayPower extends Power {
     public final String textureUrl;
     public final boolean showFirstPerson;
     public final boolean hideEntityModel;
+    public final boolean renderingPowerCompatible;
 
     public static PowerFactory<?> getFactory() {
         return new PowerFactory<>(Apugli.identifier("entity_texture_overlay"),
@@ -21,19 +22,21 @@ public class EntityTextureOverlayPower extends Power {
                         .add("texture_location", SerializableDataTypes.IDENTIFIER, null)
                         .add("url", SerializableDataTypes.STRING, null)
                         .add("show_first_person", SerializableDataTypes.BOOLEAN, false)
-                        .add("hide_player_model", SerializableDataTypes.BOOLEAN, false),
+                        .add("hide_player_model", SerializableDataTypes.BOOLEAN, false)
+                        .add("rendering_power_compatible", SerializableDataTypes.BOOLEAN,  false),
                 data ->
                         (type, player) ->
-                                new EntityTextureOverlayPower(type, player, data.getId("texture_location"), data.getString("url"), data.getBoolean("show_first_person"), data.getBoolean("hide_player_model")))
+                                new EntityTextureOverlayPower(type, player, data.getId("texture_location"), data.getString("url"), data.getBoolean("show_first_person"), data.getBoolean("hide_player_model"), data.getBoolean("rendering_power_compatible")))
                 .allowCondition();
     }
 
-    public EntityTextureOverlayPower(PowerType<?> type, LivingEntity entity, Identifier textureLocation, String textureUrl, boolean showFirstPerson, boolean hideEntityModel) {
+    public EntityTextureOverlayPower(PowerType<?> type, LivingEntity entity, Identifier textureLocation, String textureUrl, boolean showFirstPerson, boolean hideEntityModel, boolean renderingPowerCompatible) {
         super(type, entity);
         this.textureLocation = textureLocation;
         this.textureUrl = textureUrl;
         this.showFirstPerson = showFirstPerson;
         this.hideEntityModel = hideEntityModel;
+        this.renderingPowerCompatible = renderingPowerCompatible;
     }
 
     public Identifier getUrlTextureIdentifier() {